Add files via upload
If you ask anyone in the know to rank the most commonly and widely used IC, the famous 555 timer base IC would certainly be at the top of the list. The 555 - a mixed circuit composed of analog and digital circuits - integrates analogue and logical functions into an independent IC, and hence tremendously expands the application range of analog integrated circuits. The 555 is widely used in various timers, pulse generators, and oscillators.

Equipment: 
The equipment you require is as follows:
•	Laptop & Raspberry Pi 3 model Board
•	SunFounder Super Starter Kit V2.0 for Raspberry Pi 
•	NE555